fixing of the overlapping of bottom menu
