+++ pages/index.tsx
@@ -1,6 +1,6 @@
-import type { NextPage } from "next"
-import Layout from "../components/layouts/Layout"
-import MenuList from "../components/MenuList"
+import type { NextPage } from 'next'
+import Layout from '../components/layouts/Layout'
+import MenuList from '../components/MenuList'
 
 const Home: NextPage = () => {
    return (
